+++ src/controller/thread_count.rs
@@ -1,33 +1,33 @@
 pub struct ThreadCount {
-    n: f64,
-    max: f64,
+    n: f32,
+    max: f32,
 }
 
 impl ThreadCount {
     #[inline]
     pub fn new(max: i32) -> Self {
-        Self { n: max as f64, max: max as f64 }
+        Self { n: max as f32, max: max as f32 }
     }
 
     #[inline]
     fn clamp(&mut self) {
-        self.n = f64::max(1.0, f64::min(self.n, self.max));
+        self.n = f32::max(1.0, f32::min(self.n, self.max));
     }
 }
 
-impl std::ops::AddAssign<f64> for ThreadCount {
+impl std::ops::AddAssign<f32> for ThreadCount {
     #[inline]
-    fn add_assign(&mut self, rhs: f64) {
+    fn add_assign(&mut self, rhs: f32) {
         self.n += rhs;
         self.clamp();
     }
 }
 
 impl std::ops::Deref for ThreadCount {
-    type Target = f64;
+    type Target = f32;
 
     #[inline]
-    fn deref(&self) -> &f64 {
+    fn deref(&self) -> &f32 {
         &self.n
     }
 }
